passing the functions that apply gravity to the object itself instead of the view calculating to all objects
