+++ PhysicsWorld/PhysicsWorld/Presenters/View/PhysicsScene/View/PhysicsScene.swift
@@ -12,7 +12,7 @@ class PhysicsScene: UIViewController {
     var gameTimer: Timer?
     let text: String
     let label = UILabel()
-    var objects: [UIView & ObjectsProtocol] = []
+    var objects: [UIView & ObjectsProtocol & GetGravityProtocol] = []
 
     init(text: String = "") {
         self.text = text
@@ -32,7 +32,11 @@ class PhysicsScene: UIViewController {
     }
     
     func update(_ currentTime: TimeInterval){
-        calculatePhysics()
+//        calculatePhysics()
+        for n in objects {
+            n.update(deltatime: currentTime)
+            
+        }
     }
     
     func addcircle() {
@@ -40,46 +44,41 @@ class PhysicsScene: UIViewController {
         self.view.addSubview(circleView)
         self.objects.append(circleView)
     }
-    
-    func calculatePhysics() {
-        for i in objects.indices {
-            let object = objects[i]
-            
-            print(object.isDynamic)
-            guard object.isDynamic else { continue }
-            
-            let response = applyintGravity(for: object)
-            objects[i].velocity = response.velocity
-            objects[i].center = response.newPosition
-        }
-    }
+ 
+//    func calculatePhysics() {
+//        for i in objects.indices {
+//            let object = objects[i]
+//            
+//            guard object.isDynamic else { continue }
+//            
+//            let response = applyintGravity(for: object)
+//            objects[i].velocity = response.velocity
+//            objects[i].center = response.newPosition
+//        }
+//    }
 
-    func applyintGravity(for object: UIView & ObjectsProtocol) -> (velocity: CGPoint, newPosition: CGPoint) {
-        let gravity = CGPoint(x: 0, y: 9.81)
-        let deltaTime = deltaTime
-        
-        let newVelocity = CGPoint(
-            x: object.velocity.x + gravity.x * object.mass * CGFloat(deltaTime),
-            y: object.velocity.y + gravity.y * object.mass * CGFloat(deltaTime)
-        )
-        
-        let newPosition = CGPoint(
-            x: object.center.x + newVelocity.x * CGFloat(deltaTime),
-            y: object.center.y + newVelocity.y * CGFloat(deltaTime)
-        )
-        
-        if object.layer.position.y > self.view.bounds.maxY + object.layer.frame.height {
-            let initialPosition = CGPoint(x: object.layer.position.x, y: .zero - object.layer.frame.height)
-            
-            return (velocity: .zero, newPosition: initialPosition)
-        }
-        
-        return (velocity: newVelocity, newPosition: newPosition)
-        
-        
-        
-      
-    }
+//    func applyintGravity(for object: UIView & ObjectsProtocol) -> (velocity: CGPoint, newPosition: CGPoint) {
+//        let gravity = CGPoint(x: 0, y: 9.81)
+//        let deltaTime = deltaTime
+//        
+//        let newVelocity = CGPoint(
+//            x: object.velocity.x + gravity.x * object.mass * CGFloat(deltaTime),
+//            y: object.velocity.y + gravity.y * object.mass * CGFloat(deltaTime)
+//        )
+//        
+//        let newPosition = CGPoint(
+//            x: object.center.x + newVelocity.x * CGFloat(deltaTime),
+//            y: object.center.y + newVelocity.y * CGFloat(deltaTime)
+//        )
+//        
+//        if object.layer.position.y > self.view.bounds.maxY + object.layer.frame.height {
+//            let initialPosition = CGPoint(x: object.layer.position.x, y: .zero - object.layer.frame.height)
+//            
+//            return (velocity: .zero, newPosition: initialPosition)
+//        }
+//        
+//        return (velocity: newVelocity, newPosition: newPosition)
+//    }
 }
 
 #Preview {
